requirements.txt
djongo==1.3.6
web3==5.24.0
pillow==8.4.0
django-crispy-forms==1.13.0